Allow for multiple profiles
This included a slightly rewrite of how the data is stored to make it easier to have multiple profiles.

Now instead of the profiles just being inside an array, each profile is its own key within an object.
That will make it easier to find individual profiles.

Now its possible to create and switch between profiles